+++ data/concat_embedding_with_tree/classification_result/evaluation_results.csv
@@ -1,0 +1,4 @@
+dataset,accuracy,precision,recall,f1
+train,0.9877916112655627,0.9879340270516389,0.9877916112655627,0.9868798039763719
+validation,0.5848235862735621,0.594171552029243,0.5848235862735621,0.578477592166205
+test,0.5711523588553751,0.5759156085916348,0.5711523588553751,0.5654413505340247
