--- data/concat_embedding_with_tree/evaluation_results.csv
@@ -1,4 +1,0 @@
-dataset,mse,rmse,mae,r2
-train,0.14965727309929908,0.3868556230679594,0.31966196407309205,0.7932343083077398
-validation,0.6125281136127326,0.7826417530471605,0.6395562619210624,0.14741183599827745
-test,0.639858340228736,0.799911457743128,0.6489037518475781,0.13565061419028424
