+++ data/concat_embedding_with_tree/regression_result/evaluation_results.csv
@@ -1,0 +1,4 @@
+dataset,mse,rmse,mae,r2
+train,0.35037439360042805,0.59192431408114,0.48962563384999425,0.5159246033035727
+validation,0.6153169497461778,0.7844214108157539,0.6478064699613497,0.14353000816723838
+test,0.6424129596572943,0.8015066809810723,0.6529902961411286,0.13219971952309484
